% Generated by roxygen2 (4.0.0): do not edit by hand
\name{addMetaData}
\alias{addMetaData}
\title{Add meta data to CES model object}
\usage{
addMetaData(model, nest, nestString, history = "")
}
\description{
This function adds metadata to a model.  Currently this is only designed to
work with CES models. Metadata is attached as attributes (naturalCoeffs and meta)
to the object and the new object is returned from the function.
}

