% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allow_surf_acc.R
\name{Allow_surf_acc}
\alias{Allow_surf_acc}
\title{Allowable Surface Accumulation}
\usage{
Allow_surf_acc(IR)
}
\arguments{
\item{inputs}{are the intake rate}
}
\value{
outputs are the allowable surface accumulation
}
\description{
This function calculates the allowable surface accumulation for a specific soil intake rate
}
\author{
Simone Albuquerque
}
