% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cdeModel2}
\alias{cdeModel2}
\title{Fit Cobb-Douglas model with Energy}
\usage{
cdeModel2(formula, data, constrained = FALSE, ...)
}
\arguments{
  \item{formula}{a formula of the form \code{y ~ capital +
  labor + energy + time}}

  \item{data}{a data frame in which \code{formala} is
  evaluated}

  \item{constrained}{a logical indicated whether the
  coefficents are constrained. See details}

  \item{\dots}{additional arguments; currently unused.}
}
\description{
Fit Cobb-Douglas model with Energy
}
\details{
More about contranints TBA.
}

