% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cdModel2}
\alias{cdModel2}
\title{Fit a Cobb-Douglas Model}
\usage{
cdModel2(formula, data, constrained = FALSE, ...)
}
\arguments{
  \item{formala}{a formual of the form \code{y ~ capital +
  labor + time}}

  \item{data}{a data fram in which \code{formula} is
  evaluated}

  \item{contrained}{a logical indicating whether the
  parameters are contrained}
}
\value{
a CDEmodel object, which is an lm object with some additioanl attributes.
}
\description{
Fit a Cobb-Douglas Model
}

