% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_stats.R
\name{fish_stats}
\alias{fish_stats}
\title{Fish Stats}
\usage{
fish_stats(fish_vector, graph = FALSE)
}
\arguments{
\item{inputs}{are vector of fish names}
}
\value{
outputs are the total fish caught, most common fish, the rarest fish, and the total number of fish and a histogram of number of fish by type.
}
\description{
This function calculates demographic statistics on a vector of fish species
}
\author{
Simone Albuquerque & Lizzy Schattle
}
