% Generated by roxygen2 (4.0.0): do not edit by hand
\name{historicalPlot}
\alias{historicalPlot}
\title{Create a plot containing historical data}
\usage{
historicalPlot(data)
}
\arguments{
  \item{data}{a data frame containing the following
  columns: \code{iY} (indexed GDP), \code{Year} (the year
  for each row in the table), \code{iK} (indexed capital
  stock), \code{iL} (indexed labor), \code{iQ} (indexed
  thermal energy), \code{iX} (indexed exergy), \code{iU}
  (indexed useful work), and \code{Country} (2-letter
  international country code).}
}
\description{
Create a plot containing historical data
}
\details{
The function returns a figure with separate graphs for each \code{Country}
in \code{data}. The graphs are arrangd vertically with year on the horizontal axis.
You can change the vertical order of the graphs with \code{levels(data$Country)}.
}

