% Generated by roxygen2 (4.0.0): do not edit by hand
\name{linexModel2}
\alias{linexModel2}
\title{Fit LINEX models}
\usage{
linexModel2(formula, data)
}
\arguments{
  \item{formula}{a formula of the form \code{y ~ captial +
  labor + energy + time}}

  \item{data}{a data frame in which \code{formula} is
  evaluated}
}
\description{
Fit LINEX models
}

