% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boat.R
\name{mod}
\alias{mod}
\title{Easily find remainders!}
\usage{
mod(x, y)
}
\arguments{
\item{x, y}{Two integer values.}
}
\value{
Returns an integer remainder
}
\description{
This function divides x by y and finds the remainder
}
\examples{
mod(4,2) = 0
mod(7,4) = 3
}
