\name{package2-package}
\alias{package2-package}
\alias{package2}
\docType{package}
\title{
\packageTitle{package2}
}
\description{
\packageDescription{package2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{package2}
\packageIndices{package2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{package2}

Maintainer: \packageMaintainer{package2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
