% Generated by roxygen2 (4.0.0): do not edit by hand
\name{resampleFits}
\alias{resampleFits}
\title{Performs resample fits for a model}
\usage{
resampleFits(modelType = modelTypes, countryAbbrev = countryAbbrevs,
  energyType = "Q", factor = "K", method = resampleMethods, n,
  baseHistorical, rho = c(9, 2, 1, 0.43, 0.25, 0.1, -0.1, -0.5, -0.75, -0.9,
  -0.99), rho1 = c(9, 2, 1, 0.43, 0.25, 0.1, -0.1, -0.5, -0.75, -0.9, -0.99))
}
\arguments{
  \item{modelType}{one of \code{"sf"}, \code{"cd"},
  \code{"cde"}, \code{"cese-(kl)"}, \code{"cese-(kl)e"},
  \code{"cese-(le)k"}, \code{"cese-(kl)e"}, or
  \code{"linex"}.}

  \item{countryAbbrev}{the country abbreviation that you
  want to analyze.}

  \item{energyType}{one of \code{"Q"} (for thermal energy),
  \code{"X"} (for exergy), or \code{"U"} (for useful
  work).}

  \item{factor}{one of \code{"K"} (for capital stock),
  \code{"L"} (for labor), \code{"Q"} (for thermal energy),
  \code{"X"} (for exergy), or \code{"U"} (for useful
  work).}

  \item{method}{one of \code{"resample"},
  \code{"residual"}, \code{"wild"}, or \code{"debug"}}

  \item{n}{the number of resamples you want to perform}

  \item{baseHistorical}{the relative path to the directory
  containing the historical data}

  \item{rho}{a vector of \code{rho} value over which you'll
  search if doing a CES analysis}

  \item{rho1}{a vector of \code{rho1} value over which
  you'll search if doing a CES analysis}
}
\value{
a data frame containing resample data for specified countries for the given \code{modelType}, \code{energyType},
or \code{factor}.
}
\description{
This function returns a list containing models and coefficients from the resample fits.
}

