% Generated by roxygen2 (4.0.0): do not edit by hand
\name{singleFactorModel2}
\alias{singleFactorModel2}
\title{Fit a single factor model}
\usage{
singleFactorModel2(formula, data, response, factor, time, constrained = FALSE)
}
\arguments{
\item{formula}{a formula of the form \code{ y ~ factor + time }}

\item{data}{a data frame in which \code{formula} is evaluated}

\item{constrained}{a logical indicating whether the model parameters are constrained}
}
\value{
an lm object with some additional attributes
}
\description{
Fit a single factor model
}

