% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_intake.R
\name{soil_intake}
\alias{soil_intake}
\title{Soil Intake}
\usage{
soil_intake(soil_class)
}
\arguments{
\item{inputs}{are soil class: clay, silty clay, clay loam, loam, sandy loam, loamy sand, and sand}
}
\value{
outputs are IR
}
\description{
This function calculates the intake rate from a given soil class
}
\author{
Simone Albuquerque
}
