% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_classification.R
\name{species_classification}
\alias{species_classification}
\title{Species Classification}
\usage{
species_classification(population, growth)
}
\arguments{
\item{inputs}{are population size and growth rate}
}
\value{
endangered status of species
}
\description{
This function gives you a classification from a population size and growth rate
}
\author{
Lizzy Schattle
}
