% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yrs_to_extinction.R
\name{yrs_to_extinction}
\alias{yrs_to_extinction}
\title{Years to Extinction}
\usage{
yrs_to_extinction(pop_size, yr = 1)
}
\arguments{
\item{input}{population size}
}
\value{
outputs are the years to extinction for that population size if it declines by 2% annyally
}
\description{
This function calculates the years until a species is extinct with an annual decline of 2% given a population size
}
\author{
Lizzy Schattle
}
