% Generated by roxygen2 (4.0.0): do not edit by hand
\name{linexModel2}
\alias{linexModel2}
\title{Fitting LINEX models}
\usage{
linexModel2(formula, data, response, capital, labor, energy, time)
}
\arguments{
  \item{formula}{a formula of the form \code{response ~
  captial + labor + energy + time}}

  \item{response}{instead of specifying a formula,
  expressions for the components can be specified
  individually.}

  \item{capital}{instead of specifying a formula,
  expressions for the components can be specified
  individually.}

  \item{labor}{instead of specifying a formula, expressions
  for the components can be specified individually.}

  \item{energy}{instead of specifying a formula,
  expressions for the components can be specified
  individually.}

  \item{time}{instead of specifying a formula, expressions
  for the components can be specified individually.}

  \item{data}{a data frame in which \code{formula} is
  evaluated}
}
\description{
Fitting LINEX models
}

