% Generated by roxygen2 (4.0.0): do not edit by hand
\name{resampleFits}
\alias{resampleFits}
\title{Perform resample fits for a model}
\usage{
resampleFits(model, method = resampleMethods, n, save.data = FALSE, seed,
  id, ...)
}
\arguments{
  \item{model}{a model returned from \code{linexModel},
  \code{cdeModel}, \code{cesModel}, or
  \code{singleFactorModel}.  \code{model} must have been
  fit with \code{save.data = TRUE}.}

  \item{method}{one of \code{"resample"},
  \code{"residual"}, \code{"wild"}, or \code{"debug"}}

  \item{n}{the number of resamples you want to perform}

  \item{save.data}{a logical indicating whether the data
  should be saved with each model.}

  \item{seed}{a seed for \code{set.seed} to make results
  reproducible}

  \item{id}{a character vector of length 1 used as an
  identifier that will be added to the output.  This is
  convenient for marking data internally if you are doing
  resampling on many different models.}

  \item{...}{additional arguments passed to \code{fitfun}}
}
\value{
a list of length two containting a data frame of coefficients and a list of models
}
\description{
This function returns a list containing models and coefficients from the resample fits.
}

